plugin/channel/telegram_bot/requirements.txt
# Dipendenze specifiche del plugin telegram_bot.
# Installare con: uv pip install -r requirements.txt
python-telegram-bot>=21.0
pyyaml>=6.0.3
